custom_python-requirements.txt
datetime==4.3
unidecode==1.3.2
openpyxl==3.0.9
lxml==4.9.1
requests==2.28.1
pandas==1.5.3
urllib3==1.26.16
google-auth==2.21.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.3
beautifulsoup4==4.11.1
colorama
pandas-gbq